requirements.txt
# --- API + utils
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.33
httpx>=0.27,<0.28
pydantic-settings>=2.3,<3
orjson>=3.10,<4
python-dotenv>=1.0,<2
python-multipart==0.0.20
sqlmodel>=0.0.21
passlib[bcrypt]>=1.7,<2
python-jose[cryptography]>=3.3,<3.4
pyyaml>=6.0
# --- Numerics / data
# Use a NumPy range that satisfies both TensorFlow and scikit-learn wheels on Render
numpy>=1.24,<2.2
pandas==2.2.3
scikit-learn==1.5.2
statsmodels==0.14.2
duckdb==1.1.2
duckdb-engine>=0.11,<0.12
redis>=5,<6
nltk>=3.9,<4
ta==0.11.0
onnx>=1.16,<2
onnxruntime==1.17.3
scipy>=1.14,<2
pyarrow>=17,<18
discord.py>=2.4,<3

# --- Testing (optional)
pytest>=8.2,<9
pytest-asyncio>=0.23,<0.24

# --- RL stack (SB3 is PyTorch-based)
gymnasium==0.29.1
stable-baselines3==2.3.0
torch==2.3.1             # cpu wheel available for linux/windows; SB3 needs >=2.3

# --- TensorFlow (CPU builds)
# tensorflow-intel==2.15.0  # Windows-only; comment out for Linux deployment
tensorflow==2.20.0
# --- Quantum
qiskit>=1.4.1
qiskit-aer>=0.15.0
qiskit-ibm-runtime>=0.41.0
prometheus-client==0.20.0
SQLAlchemy>=2.0
